% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_CARE_region.R
\name{with_CARE_region}
\alias{with_CARE_region}
\title{with_CARE_region}
\usage{
with_CARE_region(input_data, by = "fac_id",
  using = filter(CARE_region_geodata, CARE_designation == "impact"))
}
\arguments{
\item{input_data}{data frame or tibble with column \code{fac_id}}
}
\description{
with_CARE_region
}
